Add posting and storing of page views
Add analytics.js javascript that will post to /views/record on each page
load with the current page URL.

Add analytics.less style that will hide analytics data elements from the
user.

Add RecordController and /views/record route which will accept the page
URL and store the view in the database. Record controller stores an
AnalyticsEntry in the database using an IRepository.

Update the Analytics view to add a request verification token for use in
the analytics post request.

Add IUserProvider interface and default implementation which is
responsible for returning a string that can be used to represent the
current user in the analytics stats. This can be overridden when using
different user systems.