Add simple stats viewing area
Add a permission for viewing analytics.

Add a simple area to the admin which allows users with the 'ViewAnalytics'
permission to review the analytics.

Add the Index route which simply shows all visits in chronological order
with the most recent first, displaying the URL, user identifier and visit
date/time.

Add the ByPage route which will display a break down of the visits groups
by the page URL with the number of visits for each page displayed.

Add the ByUser route which will display the visits groups by the user
identifier.

Add a queries class which holds the SQL queries use to generate the
pages.