Add Analytics.Tags feature
Add a new feature for bringing content tag info into your analytics data.

Add a new By Tag page to the analytics screen in the dashboard which will
display content visits by tags.

Add Extension method for discovering if tags feature is enabled.

Add AnalyticsSettings class which gives injectable access to current
configuration values for the analytics module. Add the TagsEnabled boolean
for easy access to this.

Add a content handler which is responsible for storing the currently
viewed content item on the work context state.

Update RecordController to store the content item id and the current item
tags if relevant.

Add tags and content item id to the AnalyticsRecord.

Refactor AnalyticsQueries again to simplify and reduce the complexity of
the analytics stats page view models.

Fix error on admin page from attempting to load classes for the analytics
image set.