Add CSV download feature
Add ability to download a CSV of the current analytics results.

Add CSVHelper nuget package for CSV writing.

Update the AdminController to add an alternative route when DownloadCsv is
true that will serve up a CSV file of the results instead of displaying
the view.

Update DTOs to adjust order that columns appear in the CSV.

Add a download CSV button to the filters partial.