Fix issue with DI when tags feature isn't enabled
When navigating to the Analytics section in the admin area, an
exception would occur. This was being caused because
`AnalyticsQueries` contained `ITagService` as a contructor argument. If
the `Orchard.Tags` module isn't enabled, this dependency couldn't be
injected causing an exception to occur. To fix this, `ITagService` has
been removed as a constructor argument and exposed as a public property.
Any methods relying on `ITagService` now have null checks that will
throw an exception.